backend/requirements.txt
asgiref==3.6.0
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.2.1
PyJWT==2.6.0
pymongo==3.12.3
pytz==2022.7
sqlparse==0.2.4
tzdata==2022.7
